Bug 1039886 - Make ProcessKill signal SIGKILL only if the child still exists r=bsmedberg

+++ ipc/chromium/src/base/process_util_posix.cc
@@ -71,16 +71,21 @@ bool KillProcess(ProcessHandle process_id, int exit_code, bool wait) {
 
   if (result && wait) {
     int tries = 60;
+    bool exited = false;
     // The process may not end immediately due to pending I/O
     while (tries-- > 0) {
       int pid = HANDLE_EINTR(waitpid(process_id, NULL, WNOHANG));
-      if (pid == process_id)
+      if (pid == process_id) {
+        exited = true;
         break;
+      }
 
       sleep(1);
     }
 
-    result = kill(process_id, SIGKILL) == 0;
+    if (!exited) {
+      result = kill(process_id, SIGKILL) == 0;
+    }
   }
 
   if (!result)
